1 添加可以滑动绿色块的做法 2 当绿条小于30时不好选中移动 此时做法进行两边留边，只滚动uiscrollview。

+++ testEasyScrollView/ViewController.m
@@ -58,7 +58,7 @@
         
         __strong __typeof(weakSelf)strongSelf = weakSelf;
         
-        strongSelf.addressLabel.text = [NSString stringWithFormat:@"%02ld:%02ld",nMbeginTimeMins/60,nMbeginTimeMins%60];
+        strongSelf.addressLabel.text = [NSString stringWithFormat:@"%02ld:%02ld",(long)(nMbeginTimeMins/60),(long)(nMbeginTimeMins%60)];
     }];
